% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsas.R
\name{converttimedate}
\alias{converttimedate}
\alias{convert_to_date}
\alias{convert_to_datetime}
\title{helper function to convert SAS date numeric to date}
\usage{
convert_to_date(x)

convert_to_datetime(x)
}
\arguments{
\item{x}{date or datetime variable}
}
\description{
helper function to convert SAS date numeric to date
}
\examples{
 # 2000-03-17
 convert_to_date(14686)
 # 2012-11-10 03:49:19
 convert_to_datetime(1668138559)
}
