% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsas.R
\name{read.sas}
\alias{read.sas}
\title{read.sas}
\usage{
read.sas(
  file,
  debug = FALSE,
  convert_dates = TRUE,
  recode = TRUE,
  rowcount = -1,
  remove_deleted = TRUE
)
}
\arguments{
\item{file}{file to read}

\item{debug}{print debug information}

\item{convert_dates}{default is TRUE}

\item{recode}{default is TRUE}

\item{rowcount}{number of rows to import (from start). negative values are
ignored}

\item{remove_deleted}{logical if deleted rows should be removed from data}
}
\description{
read.sas
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}
}
