% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readsas}
\alias{readsas}
\title{Reads SAS data files}
\usage{
readsas(filePath, debug, selectrows_, selectcols_, empty_to_na, tempstr)
}
\arguments{
\item{filePath}{The full systempath to the sas7bdat file you want to import.}

\item{debug}{print debug information}

\item{selectrows_}{integer vector of selected rows}

\item{selectcols_}{character vector of selected rows}

\item{empty_to_na}{logical convert '' to NA_character_}

\item{tempstr}{filepath used for temp output when uncompressing}
}
\description{
Reads SAS data files
}
