% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writesas}
\alias{writesas}
\title{Writes the binary SAS file}
\usage{
writesas(filePath, dat, compress, debug, bit32)
}
\arguments{
\item{filePath}{The full systempath to the sas7bdat file you want to export.}

\item{dat}{an R-Object of class data.frame.}

\item{compress}{the file}

\item{debug}{print debug information}
}
\description{
Writes the binary SAS file
}
