% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsas.R
\name{read.sas}
\alias{read.sas}
\title{read.sas}
\usage{
read.sas(
  file,
  debug = FALSE,
  convert_dates = TRUE,
  recode = TRUE,
  select.rows = NULL,
  select.cols = NULL,
  remove_deleted = TRUE,
  rownames = FALSE,
  empty_to_na = FALSE
)
}
\arguments{
\item{file}{file to read}

\item{debug}{print debug information}

\item{convert_dates}{default is \code{TRUE}}

\item{recode}{default is \code{TRUE}}

\item{select.rows}{\emph{integer.} Vector of rows to import. Minimum 0. Rows
imported are sorted.}

\item{select.cols}{\emph{character:} Vector of variables to select.}

\item{remove_deleted}{logical if deleted rows should be removed from data}

\item{rownames}{first column will be used as rowname and removed from data}

\item{empty_to_na}{logical. In SAS empty characters are missing. this option
allows to convert \code{""} to \code{NA_character_} when importing.}
}
\description{
\code{read.sas} is a general function for reading sas7bdat files.
It supports a variety of complex sas7bdat files, x86 and x64, big and small
endian, and both compression types. It has been tested with numeric and
character data, and provides helper functions for converting sas7bdat to R
types \code{Date} and \code{POSIXct}. Time variables are converted to characters.
Conversion to date variables is applied if a known date, datetime or time
format is found in the sas7bdat file. For user-defined formats, the package
provides functions to convert from sas7bdat to R.

Input files may contain deleted lines that are marked as deleted instead of
being removed from the input data. These are removed on import, if you still
need them look at \code{remove_deleted}. Formats, labels and additional file
information are available with \code{attributes()}.
}
\examples{
fl <- system.file("extdata", "cars.sas7bdat", package = "readsas")
read.sas(fl)

}
\seealso{
\link[foreign]{read.xport}
}
