% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsas.R
\name{read.sas}
\alias{read.sas}
\title{read.sas}
\usage{
read.sas(
  file,
  debug = FALSE,
  convert_dates = TRUE,
  recode = TRUE,
  select.rows = NULL,
  select.cols = NULL,
  remove_deleted = TRUE
)
}
\arguments{
\item{file}{file to read}

\item{debug}{print debug information}

\item{convert_dates}{default is TRUE}

\item{recode}{default is TRUE}

\item{select.rows}{\emph{integer.} Vector of one or two numbers. If single
value rows from 1:val are selected. If two values of a range are selected
the rows in range will be selected.}

\item{select.cols}{\emph{character:} Vector of variables to select.}

\item{remove_deleted}{logical if deleted rows should be removed from data}
}
\description{
read.sas
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}
}
