% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsas.R
\name{read.sas}
\alias{read.sas}
\title{read.sas}
\usage{
read.sas(file, debug = FALSE, convert.dates = TRUE, recode = TRUE,
  rowcount = -1)
}
\arguments{
\item{file}{file to read}

\item{debug}{print debug information}

\item{convert.dates}{default is TRUE}

\item{recode}{default is TRUE}

\item{rowcount}{number of rows to import (from start). negative values are
ignored}
}
\description{
read.sas
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}
}
