% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readsas}
\alias{readsas}
\title{Reads SAS data files}
\usage{
readsas(filePath, debug, selectrows_, selectcols_)
}
\arguments{
\item{filePath}{The full systempath to the sas7bdat file you want to import.}

\item{debug}{print debug information}

\item{selectrows_}{integer vector of selected rows}

\item{selectcols_}{character vector of selected rows}
}
\description{
Reads SAS data files
}
