% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writesas.R
\name{write.sas}
\alias{write.sas}
\title{write.sas}
\usage{
write.sas(dat, filepath, compress = 0, debug = FALSE)
}
\arguments{
\item{dat}{data frame to save}

\item{filepath}{path to save file to}

\item{compress}{option to compress file}

\item{debug}{print debug information}
}
\description{
write.sas
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}
}
